Add separate tests for entire lambda functions
