+++ database/test/test_services.py
@@ -1,0 +1,73 @@
+import inspect
+import json
+
+import psycopg2
+import pytest
+import requests
+
+from database.koodistot_loader.koodistot_loader import codes
+
+from .conftest import assert_database_is_alright, drop_hame_db
+
+
+@pytest.fixture()
+def db_manager_url(docker_ip, docker_services):
+    port = docker_services.port_for("db_manager", 8080)
+    return f"http://{docker_ip}:{port}/2015-03-31/functions/function/invocations"
+
+
+@pytest.fixture()
+def koodistot_loader_url(docker_ip, docker_services):
+    port = docker_services.port_for("koodistot_loader", 8080)
+    return f"http://{docker_ip}:{port}/2015-03-31/functions/function/invocations"
+
+
+@pytest.fixture()
+def create_db(db_manager_url, main_db_params, root_db_params):
+    payload = {
+        "event_type": 1,
+    }
+    r = requests.post(db_manager_url, data=json.dumps(payload))
+    data = r.json()
+    assert data["statusCode"] == 200, data["body"]
+    yield
+
+    drop_hame_db(main_db_params, root_db_params)
+
+
+@pytest.fixture()
+def populate_koodistot(koodistot_loader_url, main_db_params, create_db):
+    payload = {}
+    r = requests.post(koodistot_loader_url, data=json.dumps(payload))
+    data = r.json()
+    assert data["statusCode"] == 200, data["body"]
+
+
+def test_create_db(create_db, main_db_params_with_root_user):
+    """
+    Test the whole lambda endpoint
+    """
+    conn = psycopg2.connect(**main_db_params_with_root_user)
+    try:
+        with conn.cursor() as cur:
+            assert_database_is_alright(cur)
+    finally:
+        conn.close()
+
+
+def test_populate_koodistot(populate_koodistot, main_db_params):
+    """
+    Test the whole lambda endpoint
+    """
+    conn = psycopg2.connect(**main_db_params)
+    try:
+        with conn.cursor() as cur:
+            for name, value in inspect.getmembers(codes, inspect.isclass):
+                # some code tables may not have external source
+                if value.code_list_uri:
+                    print(value)
+                    cur.execute(f"SELECT count(*) FROM codes.{value.__tablename__}")
+                    code_count = cur.fetchone()[0]
+                    assert code_count > 0
+    finally:
+        conn.close()
